+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 5.2.1 / Current Snapshot
 ========================
 
+Improvements:
+
+* [OLMIS-2780](https://openlmis.atlassian.net/browse/OLMIS-2780): User form now uses minimal facilities endpoint.
+
 New functionality added in a backwards-compatible manner:
+
 * [OLMIS-2696](https://openlmis.atlassian.net/browse/OLMIS-2696): Added requisition group administration screen.
 * [OLMIS-2698](https://openlmis.atlassian.net/browse/OLMIS-2698): Added geographic zone administration screens.